requirements.txt
# --- Qiskit 0.44 Stable Release (Known Good) ---
qiskit==0.44.3
qiskit-aer==0.12.2
qiskit-machine-learning==0.6.1
qiskit-algorithms==0.2.2

# --- Standard App Dependencies ---
streamlit==1.36.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
scipy==1.13.1
matplotlib==3.8.4
plotly==5.22.0
seaborn==0.13.2
requests==2.32.3
python-dotenv==1.0.1
pylatexenc